+++ src/tests/fakes/paths.c
@@ -1,0 +1,13 @@
+#include <trx/game/paths.h>
+
+const char *GamePath_PeekResolve(
+    const GAME_DYNAMIC_PATH path, const char *const rel)
+{
+    return rel;
+}
+
+const char *GamePath_TryResolve(
+    const GAME_DYNAMIC_PATH path, const char *const rel)
+{
+    return rel;
+}